Revert "Issue #11 fixed"
