+++ .gitignore
@@ -3,4 +3,3 @@ node_modules/
 .vagrant/
 .env
 _ide_helper.php
-sftp-config.json
