--- app/CheckUrl.php
@@ -1,33 +1,0 @@
-<?php namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
-
-/**
- * The application's  url store for health check
- */
-class CheckUrl extends Model
-{
-    use SoftDeletes;
-
-    const REPORT = 1;
-
-    const ERROR = 1;
-
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array
-     */
-    protected $fillable = ['title', 'url', 'project_id', 'period', 'is_report'];
-
-    /**
-     * Belongs to relationship
-     *
-     * @return Project
-     */
-    public function project()
-    {
-        return $this->belongsTo('App\Project');
-    }
-}
