+++ app/Console/Kernel.php
@@ -30,8 +30,7 @@ class Kernel extends ConsoleKernel
      * @var array
      */
     protected $commands = [
-        'App\Console\Commands\CheckHeartbeats',
-        'App\Console\Commands\CheckUrl'
+        'App\Console\Commands\CheckHeartbeats'
     ];
 
     /**
@@ -43,9 +42,6 @@ class Kernel extends ConsoleKernel
     protected function schedule(Schedule $schedule)
     {
         $schedule->command('heartbeat:check')
-                 ->everyFiveMinutes()
-                 ->withoutOverlapping();
-        $schedule->command('checkurl')
                  ->everyFiveMinutes()
                  ->withoutOverlapping();
     }
