--- app/ProjectRelationModel.php
@@ -1,105 +1,0 @@
-<?php namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-
-/**
- * Project model's relations
- */
-abstract class ProjectRelationModel extends Model
-{
-    /**
-     * Belongs to relationship
-     *
-     * @return Group
-     */
-    public function group()
-    {
-        return $this->belongsTo('App\Group');
-    }
-
-    /**
-     * Has many relationship
-     *
-     * @return Server
-     */
-    public function servers()
-    {
-        return $this->hasMany('App\Server')->orderBy('name');
-    }
-
-    /**
-     * Has many relationship
-     *
-     * @return Heartbeat
-     */
-    public function heartbeats()
-    {
-        return $this->hasMany('App\Heartbeat')->orderBy('name');
-    }
-
-    /**
-     * Has many relationship
-     *
-     * @return Notification
-     */
-    public function notifications()
-    {
-        return $this->hasMany('App\Notification')->orderBy('name');
-    }
-
-    /**
-     * Has many relationship
-     *
-     * @return Deployment
-     */
-    public function deployments()
-    {
-        return $this->hasMany('App\Deployment')->orderBy('started_at', 'DESC');
-    }
-
-    /**
-     * Has many relationship
-     *
-     * @return Command
-     */
-    public function commands()
-    {
-        return $this->hasMany('App\Command');
-    }
-
-    /**
-     * Has many relationship
-     * @return SharedFile
-     */
-    public function shareFiles()
-    {
-        return $this->hasMany('App\SharedFile');
-    }
-
-    /**
-     * Has many relationship to project file
-     * @return ProjectFile
-     */
-    public function projectFiles()
-    {
-        return $this->hasMany('App\ProjectFile');
-    }
-
-    /**
-     * Has many relationship
-     * @return SharedFile
-     */
-    public function notifyEmails()
-    {
-        return $this->hasMany('App\NotifyEmail');
-    }
-
-    /**
-     * Has many urls to check
-     * @return CheckUrl
-     */
-    public function checkUrls()
-    {
-        return $this->hasMany('App\CheckUrl');
-    }
-}
