+++ app/Providers/RouteServiceProvider.php
@@ -42,7 +42,6 @@ class RouteServiceProvider extends ServiceProvider
         $router->model('shared-files', 'App\SharedFile');
         $router->model('project-file', 'App\ProjectFile');
         $router->model('notify-email', 'App\NotifyEmail');
-        $router->model('check-url', 'App\CheckUrl');
     }
 
     /**
