+++ gulpfile.js
@@ -61,7 +61,6 @@ elixir(function(mix) {
         'notifyEmails.js',
         'shareFiles.js',
         'projectFiles.js',
-        'checkUrls.js',
         'deployment.js',
         'commands.js',
         'users.js',
