--- resources/lang/en/checkUrls.php
@@ -1,20 +1,0 @@
-<?php
-
-return [
-
-    'label'              => 'Check Links',
-    'create'             => 'Add a new url',
-    'edit'               => 'Edit the url',
-    'view'               => 'View the url',
-    'none'               => 'You can add urls which should be requested periodically after a deployment finished.',
-    'title'              => 'Title',
-    'titleTip'           => 'The using of this url',
-    'url'                => 'URL',
-    'period'             => 'Period',
-    'is_report'          => 'Report when failed',
-    'length'             => 'Minutes',
-    'warning'            => 'The link could not be saved, please check the form below.',
-    'urlExample'         => 'A valid url link, must begin with http or https',
-    'last_status'        => 'Last Status'
-
-];
