+++ resources/lang/en/deployments.php
@@ -27,6 +27,7 @@ return [
     'failed'           => 'Failed',
     'running'          => 'Running',
     'cancelled'        => 'Cancelled',
+    'loading'          => 'Loading',
     'reason'           => 'Reason for deployment',
     'source'           => 'Source',
     'default'          => 'Default branch (:branch)',
